test(runtime): modernize concurrency conformance checks
Record the remaining runtime, actor, supervision, sync, time, trace, IO, and security-key changes as one runtime/concurrency cleanup slice after protocol and ATP-specific work were separated.

Actor, GenServer, Monitor, and Supervision conformance paths now report failure collections through assert! messages instead of ad hoc panic branches, and their helper test modules carry explicit clippy allowances where vector push setup is intentional. Panic-triggering mock handlers use assert! conditions so the intended panic points stay visible to the test harness.

Runtime metamorphic and reactor tests were tightened with range contains checks, boundary-value preservation assertions, simpler iteration, clearer epoch arithmetic, assert!-based critical failure reporting, and registration/io-driver conformance cleanups. The sync/time/trace tests receive the same style of deterministic assertion cleanup, and the runtime scheduler/rwlock/sync integration tests stay with the primitives they validate.

The small supporting changes in cancel symbol metamorphic coverage, Cx registry behavior, IO seek extension, security key handling, deterministic hash utilities, and runtime state metamorphic coverage are grouped here because they all sit under the same cancel-correct runtime invariant surface rather than the ATP or broker protocol surfaces.

Staging note: .ntm runtime artifacts were intentionally excluded. Validation note: git diff --cached --check passed for this staged grouping; no file contents were edited during commit preparation.

+++ src/runtime/epoch_tracker.rs
@@ -1287,7 +1287,7 @@ mod tests {
                     ModuleId::TaskTable,
                     EpochId::new(to - 1),
                     EpochId::new(to),
-                    Time::from_nanos(1000 + (to as u64) * 100),
+                    Time::from_nanos(1000 + to * 100),
                 );
             }
             tracker.notify_epoch_transition(
